+++ inc/crn_decomp.h
@@ -44,8 +44,8 @@ namespace crnd
    typedef unsigned int       uint;
    typedef signed int         int32;
    #if defined(__APPLE__)
-      typedef unsigned long         uint64;
-      typedef long                  int64;
+      typedef unsigned long long    uint64;
+      typedef long long             int64;
    #elif defined(__GNUC__)
       typedef unsigned long long    uint64;
       typedef long long             int64;
